feat: add endpoints for all tables

+++ FinalProject/api/models/menu_items.py
@@ -1,6 +1,6 @@
 from decimal import Decimal
 
-from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME, Double
+from sqlalchemy import Column, ForeignKey, Integer, String, DECIMAL, DATETIME, Double, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..dependencies.database import Base
@@ -11,7 +11,7 @@ class MenuItem(Base):
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String(50))
     description = Column(String(200))
-    price = Column(DECIMAL)
+    price = Column(Float)
     calories = Column(Integer)
     food_category = Column(String(50))
